Implement console prototype of Torrent Tracker
Implement AbstractServer, which implements Server interface and handles
clients with clients handler given by HandlerFactory.

Implement TorrentServer, based on AbstractServer with TrackerClientHandler
as client handler.

Implement TrackerClientImpl, which allows to connect with TorrentServer
and send requests to it.

Implement PeerToPeerConnection, which unites AbstractServer and
PeerToPeerClient in itself. This class allows to communicate to download
file.

Implement TorrentClient, which unites TrackerClientImpl and
PeerToPeerConnection and provides easy interface for user.

Add console prototype (.main() in TorrentServer and in TorrentClient).

Add tests for list, upload and download client functions.